requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
pandas
numpy
spotipy
python-dotenv
tiktoken
tqdm
scikit-learn
rank-bm25
sentence-transformers